feat:CEXT-1537

+++ src/generator-create-api-mesh.js
@@ -32,19 +32,31 @@ class ApiMeshCreateGenerator extends Generator {
         this.log(meshPluginName + ' is not installed. Installing...')
         await this.spawnCommand('aio', ['plugins', 'install', meshPluginName])
       }
-      const output = (await this.spawnCommand('aio', ['api-mesh', 'create', '-c', this.options['template-folder'] + '/conf/mesh.json', '--json'], { stdio: [process.stderr] })).stdout
 
-      this.props.meshConfig = JSON.parse(output.substring(output.indexOf('{'), output.lastIndexOf('}') + 1))
-      const dotenvFile = this.destinationPath(constants.dotenvFilename)
-      const vars = [{ name: 'MESH_ID', value: this.props.meshConfig.mesh.meshId }, { name: 'MESH_API_KEY', value: this.props.meshConfig.adobeIdIntegrationsForWorkspace.apiKey }]
-      const content = `${vars.map(v => `${v.name}=${v.value}`).join(EOL)}${EOL}`
-      /* istanbul ignore next */
-      if (this.fs.exists(dotenvFile)) {
+      //Check if mesh already exists in the selected org/projrct/workspace
+      const checkIfMeshExists = (await this.spawnCommand('aio', ['api-mesh', 'get'], { stdio: 'ignore'}))
+
+      //If mesh does not exist then return the exitCode
+      if(!checkIfMeshExists.exitCode){
+        const output = (await this.spawnCommand('aio', ['api-mesh', 'create', '-c', this.options['template-folder'] + '/conf/mesh.json', '--json'], { stdio: [process.stderr] })).stdout
+
+        this.props.meshConfig = JSON.parse(output.substring(output.indexOf('{'), output.lastIndexOf('}') + 1))
+        const dotenvFile = this.destinationPath(constants.dotenvFilename)
+        const vars = [{ name: 'MESH_ID', value: this.props.meshConfig.mesh.meshId }, { name: 'MESH_API_KEY', value: this.props.meshConfig.adobeIdIntegrationsForWorkspace.apiKey }]
+        const content = `${vars.map(v => `${v.name}=${v.value}`).join(EOL)}${EOL}`
         /* istanbul ignore next */
-        this.fs.append(dotenvFile, content)
-      } else {
-        this.fs.write(dotenvFile, content)
+        if (this.fs.exists(dotenvFile)) {
+          /* istanbul ignore next */
+          this.fs.append(dotenvFile, content)
+        } else {
+          this.fs.write(dotenvFile, content)
       }
+      }
+      else{
+        throw new Error('Selected org, project and workspace already has a mesh. Please delete the mesh to create a sample mesh using aio app')
+      }
+
+      
     } catch (error) {
       this.log(error)
       throw new Error('Unable to create a mesh. Run "aio plugins" and check if "@adobe/aio-cli-plugin-api-mesh" plugin is installed. Make sure you\'re using the latest version.')
